feat: all encoders got integrated

+++ src/shared/connectivity/sequential.rs
@@ -1,7 +1,7 @@
 pub(crate) const INDEX_SIZE_SLOT_SELECTOR: [usize;3] = [8,16,32];
-pub(crate)const INDEX_SIZE_SLOT: usize = 2;
-pub(crate) const NUM_POINTS_SLOT: usize = 32;
-pub(crate) const NUM_FACES_SLOT: usize = 32;
+pub(crate)const INDEX_SIZE_SLOT: u8 = 2;
+pub(crate) const NUM_POINTS_SLOT: u8 = 32;
+pub(crate) const NUM_FACES_SLOT: u8 = 32;
 
 pub(crate) fn index_size_from_vertex_count(vertex_count: usize) -> Result<usize, Err> {
     match vertex_count {
